most: make sure we copy in qemu static binary
Because the new build host is debian, we need to make sure that
the qemu binaries are copied in, systemd doesn't happen transparently yet.
Next Debian stable release should make this commit obsolete.

Also don't mkdir the usr/bin directory, we actually want it to error out
if it doesn't exist.