pinebook: Initial work towards the image
It boots properly, but due to the 4.19 kernel, the left side USB doesn't
work, nor does audio, or bluetooth.  Better support will arrive when the
5.1 kernel is introduced in Debian.

Wireless driver is out of tree, and is in /usr/src but the service
doesn't actually seem to run "dkms install ." like it should, though the
service itself starts and runs without issue, so users need to go into
/usr/src/rtl8723cs and run `dkms install .` in there in order to build
and install the working wireless driver.  External USB devices work, as
long as they are used in the right side port.